+++ backend/src/main/java/com/utephonehub/dao/auth/PasswordResetTokenDao.java
@@ -8,8 +8,8 @@ import com.utephonehub.model.auth.PasswordResetToken;
 public interface PasswordResetTokenDao {
     PasswordResetToken create(PasswordResetToken token) throws SQLException;
     Optional<PasswordResetToken> findByToken(String token) throws SQLException;
-    Optional<PasswordResetToken> findByUserId(long userId) throws SQLException;
-    int delete(long id) throws SQLException;
-    int deleteByUserId(long userId) throws SQLException;
+    Optional<PasswordResetToken> findByUserId(int userId) throws SQLException;
+    int delete(int id) throws SQLException;
+    int deleteByUserId(int userId) throws SQLException;
     int deleteExpiredTokens() throws SQLException;
 }
