+++ backend/src/main/java/com/utephonehub/dao/cart/CartDAO.java
@@ -11,7 +11,7 @@ import java.util.Optional;
  * DAO interface for Cart entity
  * Module M05 - Shopping Cart
  */
-public interface CartDAO extends GenericDAO<Cart, Long> {
+public interface CartDAO extends GenericDAO<Cart, Integer> {
     
     /**
      * Find cart by user ID
@@ -19,7 +19,7 @@ public interface CartDAO extends GenericDAO<Cart, Long> {
      * @return Optional containing Cart if found, empty otherwise
      * @throws SQLException if database error occurs
      */
-    Optional<Cart> findByUserId(long userId) throws SQLException;
+    Optional<Cart> findByUserId(int userId) throws SQLException;
     
     /**
      * Find cart by user ID within a transaction
@@ -28,7 +28,7 @@ public interface CartDAO extends GenericDAO<Cart, Long> {
      * @return Optional containing Cart if found, empty otherwise
      * @throws SQLException if database error occurs
      */
-    Optional<Cart> findByUserId(long userId, Connection conn) throws SQLException;
+    Optional<Cart> findByUserId(int userId, Connection conn) throws SQLException;
     
     /**
      * Create or get existing cart for user
@@ -36,7 +36,7 @@ public interface CartDAO extends GenericDAO<Cart, Long> {
      * @return Cart (existing or newly created)
      * @throws SQLException if database error occurs
      */
-    Cart getOrCreateCartForUser(long userId) throws SQLException;
+    Cart getOrCreateCartForUser(int userId) throws SQLException;
     
     /**
      * Create or get existing cart for user within a transaction
@@ -45,14 +45,14 @@ public interface CartDAO extends GenericDAO<Cart, Long> {
      * @return Cart (existing or newly created)
      * @throws SQLException if database error occurs
      */
-    Cart getOrCreateCartForUser(long userId, Connection conn) throws SQLException;
+    Cart getOrCreateCartForUser(int userId, Connection conn) throws SQLException;
     
     /**
      * Update cart's updated_at timestamp
      * @param cartId Cart ID to update
      * @throws SQLException if database error occurs
      */
-    void updateTimestamp(long cartId) throws SQLException;
+    void updateTimestamp(int cartId) throws SQLException;
     
     /**
      * Update cart's updated_at timestamp within a transaction
@@ -60,5 +60,5 @@ public interface CartDAO extends GenericDAO<Cart, Long> {
      * @param conn Database connection for transaction
      * @throws SQLException if database error occurs
      */
-    void updateTimestamp(long cartId, Connection conn) throws SQLException;
+    void updateTimestamp(int cartId, Connection conn) throws SQLException;
 }
