+++ backend/src/main/java/com/utephonehub/dao/user/UserDao.java
@@ -6,12 +6,12 @@ import java.util.Optional;
 import com.utephonehub.model.user.User;
 
 public interface UserDao {
-    Optional<User> findById(long id) throws SQLException;
+    Optional<User> findById(int id) throws SQLException;
     Optional<User> findByEmail(String email) throws SQLException;
     boolean existsEmail(String email) throws SQLException;
     User create(User user) throws SQLException;
-    int updateProfile(long id, String fullName, String phoneNumber) throws SQLException;
-    int updatePassword(long id, String passwordHash) throws SQLException;
+    int updateProfile(int id, String fullName, String phoneNumber) throws SQLException;
+    int updatePassword(int id, String passwordHash) throws SQLException;
 }
